# resources/specifications/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Laravel Calendar Service API",
    "version": "0.1.0",
    "description": "RESTful API specification for managing calendars, working hours, blocked times, availability, and events."
  },
  "servers": [
    {
      "url": "http://localhost/api",
      "description": "Local Sail API base path"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/events/{eventId}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get an event",
        "operationId": "getEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          },
          {
            "$ref": "#/components/parameters/EventRevision"
          },
          {
            "$ref": "#/components/parameters/IncludeEventRelations"
          }
        ],
        "responses": {
          "200": {
            "description": "Event resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResourceResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Events"
        ],
        "summary": "Update an event",
        "operationId": "updateEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResourceResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "summary": "Delete an event",
        "operationId": "deleteEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          }
        ],
        "responses": {
          "204": {
            "description": "Event deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "EventId": {
        "name": "eventId",
        "in": "path",
        "required": true,
        "description": "Event identifier.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "EventRevision": {
        "name": "revision",
        "in": "query",
        "description": "Return the specified historical event revision. Omit this filter to return the latest revision.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "example": 2
        }
      },
...
      "IncludeEventRelations": {
        "name": "include",
        "in": "query",
        "description": "Comma-separated event-related data to embed. Supported values: assignees, sources.",
        "schema": {
          "type": "string",
          "example": "assignees,sources"
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication is required or has failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "The request conflicts with the current resource state, typically due to overlapping events.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ValidationError": {
        "description": "The request payload failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "The requested resource was not found."
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message",
          "errors"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "The given data was invalid."
          },
          "errors": {
            "type": "object",
            "description": "Validation errors grouped by field path.",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "ends_at": [
                "The ends_at field must be a date after starts_at."
              ]
            }
          }
        }
      },
...
      "EventStatus": {
        "type": "string",
        "enum": [
          "draft",
          "pending",
          "tentative",
          "approved",
          "confirmed",
          "processing",
          "processed",
          "completed",
          "cancelled",
          "rejected",
          "no_show",
          "expired",
          "deleted",
          "archived"
        ],
        "example": "approved"
      },
      "EventType": {
        "type": "string",
        "enum": [
          "single",
          "recurring_master",
          "recurring_instance",
          "recurring_exception",
          "hold",
          "availability_block",
          "external_sync"
        ],
        "example": "single"
      },
      "DayOfWeek": {
        "type": "string",
        "enum": [
          "monday",
          "tuesday",
          "wednesday",
          "thursday",
          "friday",
          "saturday",
          "sunday"
        ],
        "example": "monday"
      },
      "UserReference": {
        "type": "object",
        "description": "Normalized user reference with a resource kind and UUID identifier.",
        "additionalProperties": false,
        "required": [
          "kind",
          "id"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Resource kind of the user reference.",
            "maxLength": 100,
            "example": "user"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Primary key of the related user.",
            "example": "d2f5cd63-f5e9-4d57-b733-37ea6ce0d6c6"
          }
        }
      },
      "SourceReference": {
        "type": "object",
        "description": "External or cross-system source reference linked to the event.",
        "additionalProperties": false,
        "required": [
          "kind",
          "provider",
          "id"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Source kind, such as google_calendar_event, microsoft_graph_event, imported_csv_row, or crm_booking.",
            "maxLength": 100,
            "example": "google_calendar_event"
          },
          "provider": {
            "type": "string",
            "description": "Name of the upstream provider or source system.",
            "maxLength": 100,
            "example": "google"
          },
          "id": {
            "type": "string",
            "description": "Stable identifier from the upstream provider.",
            "maxLength": 255,
            "example": "8l2m4a3d2@example.com"
          },
          "synced_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the most recent successful sync for this source reference.",
            "example": "2026-06-27T18:45:00+02:00"
          },
          "metadata": {
            "type": "object",
            "description": "Optional provider-specific attributes related to the source reference.",
            "additionalProperties": true,
            "example": {
              "calendar_name": "Engineering",
              "import_batch": "batch-42"
            }
          }
        }
      },
      "Location": {
        "type": "object",
        "description": "Optional physical or virtual location attached to the event.",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "description": "Location classification.",
            "enum": [
              "in_person",
              "virtual",
              "hybrid"
            ],
            "example": "hybrid"
          },
          "label": {
            "type": "string",
            "description": "Short display label for the location.",
            "maxLength": 255,
            "example": "Berlin HQ / Zoom"
          },
          "address": {
            "type": "string",
            "description": "Postal address or room details for in-person events.",
            "example": "Room 3.14, Berlin HQ, Friedrichstrasse 10"
          },
          "meeting_url": {
            "type": "string",
            "format": "uri",
            "description": "Join URL for a virtual meeting.",
            "example": "https://meet.example.com/booking/q3-planning"
          }
        }
      },
      "Recurrence": {
        "type": "object",
        "description": "Recurrence rules used when the event belongs to a repeating series.",
        "additionalProperties": false,
        "required": [
          "frequency",
          "interval"
        ],
        "properties": {
          "frequency": {
            "type": "string",
            "description": "Base recurrence cadence.",
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "yearly"
            ],
            "example": "weekly"
          },
          "interval": {
            "type": "integer",
            "description": "Number of frequency units between each occurrence.",
            "minimum": 1,
            "example": 1
          },
          "count": {
            "type": "integer",
            "description": "Optional maximum number of generated occurrences.",
            "minimum": 1,
            "example": 12
          },
          "until": {
            "type": "string",
            "format": "date-time",
            "description": "Optional end boundary for the recurring series.",
            "example": "2026-09-30T23:59:59+02:00"
          },
          "by_week_days": {
            "type": "array",
            "description": "Days of the week used for weekly recurrence patterns.",
            "items": {
              "$ref": "#/components/schemas/DayOfWeek"
            },
            "example": [
              "monday",
              "wednesday"
            ]
          }
        }
      },
...
      "Event": {
        "type": "object",
        "description": "An event resource that represents a scheduled record on a calendar, including revision metadata, recurrence context, ownership, and operational metadata.",
        "additionalProperties": false,
        "required": [
          "id",
          "calendar_id",
          "sources",
          "title",
          "status",
          "type",
          "timezone",
          "starts_at",
          "ends_at",
          "created_at",
          "created_by"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Primary key of the event.",
            "example": "1b6c0d93-cfd7-4126-9cc2-b3c9394eb6d8"
          },
          "calendar_id": {
            "type": "string",
            "format": "uuid",
            "description": "Primary key of the calendar that owns this event.",
            "example": "6c6b8f0a-6cf6-4db4-9f86-3a2f51571310"
          },
          "sources": {
            "type": "array",
            "description": "Upstream or cross-system source references connected to this event.",
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "default": []
          },
          "parent_event_id": {
            "type": "string",
            "format": "uuid",
            "description": "Optional parent event identifier when this record is derived from another event.",
            "example": "0ba4cfe0-6acb-4d33-8e47-a31264d43d0d"
          },
          "series_master_event_id": {
            "type": "string",
            "format": "uuid",
            "description": "Optional master-series event identifier for recurring instances or exceptions.",
            "example": "314e0b0d-b0d7-4ae7-8808-d6b721ae2b3d"
          },
          "event_number": {
            "type": "string",
            "description": "Human-friendly event reference for support, invoices, or operator workflows.",
            "maxLength": 100,
            "example": "BK-2026-000245"
          },
          "title": {
            "type": "string",
            "description": "Short human-readable title of the event.",
            "maxLength": 255,
            "example": "Quarterly planning session"
          },
          "description": {
            "type": "string",
            "description": "Optional long-form notes or business context for the event.",
            "example": "Planning session for Q3 roadmap with engineering and product leads."
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "revision": {
            "type": "integer",
            "description": "Automatically assigned revision number for this version of the event. Each change creates a new revision; use it to retrieve a historical version.",
            "minimum": 1,
            "example": 2
          },
          "type": {
            "$ref": "#/components/schemas/EventType"
          },
          "timezone": {
            "type": "string",
            "description": "IANA timezone used to interpret the event schedule.",
            "example": "Europe/Berlin"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "description": "Inclusive start timestamp of the event in ISO 8601 format.",
            "example": "2026-07-03T09:00:00+02:00"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "description": "Exclusive end timestamp of the event in ISO 8601 format.",
            "example": "2026-07-03T10:30:00+02:00"
          },
          "all_day": {
            "type": "boolean",
            "description": "Whether the event spans the full local calendar day rather than a time range.",
            "default": false,
            "example": false
          },
          "blocks_availability": {
            "type": "boolean",
            "description": "Whether the event prevents another availability-blocking event from overlapping its time range in the same calendar.",
            "default": true,
            "example": true
          },
          "duration_minutes": {
            "type": "integer",
            "description": "Total booked duration in minutes for faster filtering and reporting.",
            "minimum": 1,
            "example": 90
          },
          "buffer_before_minutes": {
            "type": "integer",
            "description": "Optional preparation time reserved immediately before the event.",
            "minimum": 0,
            "example": 15
          },
          "buffer_after_minutes": {
            "type": "integer",
            "description": "Optional wrap-up time reserved immediately after the event.",
            "minimum": 0,
            "example": 15
          },
          "assignees": {
            "type": "array",
            "description": "Users responsible for owning or operating the event.",
            "items": {
              "$ref": "#/components/schemas/UserReference"
            }
          },
          "created_by": {
            "$ref": "#/components/schemas/UserReference"
          },
          "updated_by": {
            "$ref": "#/components/schemas/UserReference"
          },
          "updated_reason": {
            "type": "string",
            "description": "Optional explanation for the latest event change.",
            "maxLength": 500,
            "example": "Rescheduled after the customer requested a later start time."
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "recurrence": {
            "$ref": "#/components/schemas/Recurrence"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the event was created.",
            "example": "2026-06-27T17:32:11+02:00"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the event was last updated.",
            "example": "2026-06-27T18:10:02+02:00"
          },
          "notes": {
            "type": "string",
            "description": "Internal operational notes that should not necessarily be displayed to end users.",
            "example": "Room access code will be shared 15 minutes before the event starts."
          },
          "metadata": {
            "type": "object",
            "description": "Flexible extension point for provider-specific or domain-specific attributes.",
            "additionalProperties": true,
            "example": {
              "priority": "high",
              "department": "engineering",
              "requires_projector": true
            }
          }
        }
      },
...
      "UpdateEventRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            }
          },
          "parent_event_id": {
            "type": "string",
            "format": "uuid"
          },
          "series_master_event_id": {
            "type": "string",
            "format": "uuid"
          },
          "event_number": {
            "type": "string",
            "maxLength": 100
          },
          "title": {
            "type": "string",
            "maxLength": 255
          },
          "description": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "updated_reason": {
            "type": "string",
            "maxLength": 500
          },
          "type": {
            "$ref": "#/components/schemas/EventType"
          },
          "timezone": {
            "type": "string"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time"
          },
          "all_day": {
            "type": "boolean"
          },
          "blocks_availability": {
            "type": "boolean"
          },
          "buffer_before_minutes": {
            "type": "integer",
            "minimum": 0
          },
          "buffer_after_minutes": {
            "type": "integer",
            "minimum": 0
          },
          "assignees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserReference"
            }
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "recurrence": {
            "$ref": "#/components/schemas/Recurrence"
          },
          "notes": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "EventResourceResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Event"
          }
        }
      }
    }
  }
}